Move setup account request implementation out of index.js file
