+++ test/setup-account-request/setup-account-request.js
@@ -16,7 +16,7 @@ describe('setupAccountRequest called with authorisationServerId', () => {
     process.env.ASPSP_AUTH_SERVER_CLIENT_SECRET = clientSecret;
     postTokenStub = sinon.stub().returns({ access_token: accessToken });
     setupAccountRequest = proxyquire(  // eslint-disable-line
-      '../../app/setup-account-request',
+      '../../app/setup-account-request/setup-account-request',
       { './obtain-access-token': { postToken: postTokenStub } },
     ).setupAccountRequest;
   });
